feature/validation-handler-revamp, deleted obsolete files, renamed telegram handlers folders
