+++ source/group_eq_bot/drivers/local/text/base.py
