+++ source/group_eq_bot/drivers/models/supported_extensions.py
