+++ source/group_eq_bot/interfaces/telegram_event_handlers/conversation_update/helpers.py
@@ -9,13 +9,13 @@ from telegram import Update as TelegramEvent
 from telegram.ext import ContextTypes
 
 from utilities.internal_logger.logger import logger
-from driver.local.reader import Reader
+from drivers.local.reader import Reader
 from interfaces.telegram_event_validator.validator import EventValidator
 from interfaces.telegram_event_processors.public.message import MessageEventProcessor
 
 # Fetch bot configuration with hydra compose api
 # https://hydra.cc/docs/advanced/compose_api/
-initialize(version_base="1.2", config_path="../../../../configurations", job_name="conversation_validation_handler")
+initialize(version_base="1.2", config_path="../../../configurations", job_name="conversation_validation_handler")
 configurations = compose(config_name="configuration")
 GlobalHydra.instance().clear()
