+++ source/group_eq_bot/interfaces/telegram_event_handlers/message_update/document/handler.py
